requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Data processing
opencv-python>=4.8.0
scikit-learn>=1.3.0
pillow>=10.0.0

# Training utilities
tensorboard>=2.13.0
tqdm>=4.65.0

# Metrics
editdistance>=0.6.0

# API & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
matplotlib>=3.7.0
